refactor all the exporter functionality into the exporters module
