+++ src/main/python/afp_cli/cli.py
@@ -23,70 +23,22 @@ Options:
 from __future__ import print_function, absolute_import, division
 import getpass
 import os
-import subprocess
-import sys
-import tempfile
 
 from docopt import docopt
 from afp_cli import AWSFederationClientCmd, aws_credentials_file
-from .cli_functions import (format_aws_credentials,
-                            get_default_afp_server,
+from .cli_functions import (get_default_afp_server,
                             get_aws_credentials,
                             get_first_role,
-                            format_account_and_role_list,
                             )
+from .exporters import (format_aws_credentials,
+                        format_account_and_role_list,
+                        start_subshell,
+                        start_subcmd,
+                        )
 from afp_cli.log import error, debug
 from .password_providers import get_password
 from . import log
 from .config import load_config
-
-
-RC_SCRIPT_TEMPLATE = """
-# Pretend to be an interactive, non-login shell
-for file in /etc/bash.bashrc ~/.bashrc; do
-    [ -f "$file" ] && . "$file"
-done
-
-function afp_minutes_left {{
-    if ((SECONDS >= {valid_seconds})) ; then
-        echo EXPIRED
-    else
-        echo $((({valid_seconds}-SECONDS)/60)) Min
-    fi
-}}
-
-PS1="(AWS {account}/{role} \\$(afp_minutes_left)) $PS1"
-"""
-
-BATCH_FILE_TEMPLATE = """
-@echo off
-set PROMPT=$C AWS {account}/{role} $F
-"""
-
-
-def start_subshell(aws_credentials, role, account):
-    print("Press CTRL+D to exit.")
-    rc_script = tempfile.NamedTemporaryFile(mode='w')
-    rc_script.write(RC_SCRIPT_TEMPLATE.format(role=role, account=account,
-                                              valid_seconds=aws_credentials['AWS_VALID_SECONDS']))
-    rc_script.write(format_aws_credentials(aws_credentials, prefix='export '))
-    rc_script.flush()
-    subprocess.call(
-        ["bash", "--rcfile", rc_script.name],
-        stdout=sys.stdout, stderr=sys.stderr, stdin=sys.stdin)
-    print("Left AFP subshell.")
-
-
-def start_subcmd(aws_credentials, role, account):
-    batch_file = tempfile.NamedTemporaryFile(suffix=".bat", delete=False)
-    batch_file.write(BATCH_FILE_TEMPLATE.format(role=role, account=account))
-    batch_file.write(format_aws_credentials(aws_credentials, prefix='set '))
-    batch_file.flush()
-    batch_file.close()
-    subprocess.call(
-        ["cmd", "/K", batch_file.name])
-    print("Left AFP subcmd.")
-    os.unlink(batch_file.name)
 
 
 def main():
